+++ ofdrw-converter/src/test/java/HelloWorld.java
@@ -20,13 +20,13 @@ public class HelloWorld {
     public static void main(String[] args) {
 
         FontLoader.getInstance()
-            .addAliasMapping(null, "小标宋体", "方正小标宋简体", "方正小标宋简体")
-            .addAliasMapping(null, "KaiTi_GB2312", "楷体", "楷体")
+            .addAliasMapping("小标宋体", "方正小标宋简体")
+            .addAliasMapping("KaiTi_GB2312", "楷体")
 
-            .addSimilarFontReplaceRegexMapping(null, ".*Kai.*", null, "楷体")
-            .addSimilarFontReplaceRegexMapping(null, ".*SimSun.*", null, "SimSun")
-            .addSimilarFontReplaceRegexMapping(null, ".*Song.*", null, "宋体")
-            .addSimilarFontReplaceRegexMapping(null, ".*MinionPro.*", null, "SimSun");
+            .addSimilarFontReplaceRegexMapping( ".*Kai.*", "楷体")
+            .addSimilarFontReplaceRegexMapping( ".*SimSun.*", "SimSun")
+            .addSimilarFontReplaceRegexMapping( ".*Song.*",  "宋体")
+            .addSimilarFontReplaceRegexMapping( ".*MinionPro.*", "SimSun");
 
         FontLoader.setSimilarFontReplace(true);